UserService interface로 변경
